+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./management/client";
 export * from "./delivery/client";
+export { Filter, FieldFilter } from "./delivery/getDataList";
 export { ClientConfig } from "./config/config";