+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.3.0
+
+-   (feature) Add `EventEnvelope` as scalar
+-   (feature) Add `@aggregateEvents`
+-   (feature) Add `@index`
+-   (feature) Add `@webhook`
+-   (feature) Add filter for `getDataList` of delivery client
+
 # v0.2.0
 
 -   (bc) Improve object directives